clients/web: remove old login pages and consolidate components
