+++ clients/apps/web/src/app/(main)/auth/page.tsx
@@ -2,7 +2,7 @@ import { PolarLogotype } from '@/components/Layout/Public/PolarLogotype'
 import { CONFIG } from '@/utils/config'
 import { Metadata } from 'next'
 import { cookies } from 'next/headers'
-import Auth from './Auth'
+import Auth from '@/components/Auth/Auth'
 import { getServerSideAPI } from '@/utils/client/serverside'
 import { checkAuthenticationSession } from '@/utils/auth'
 import { schemas } from '@polar-sh/client'
