+++ clients/apps/web/src/components/Auth/AppleLoginButton.tsx
@@ -1,42 +1,54 @@
+import { useAuthSessionStart } from '@/hooks'
 import { usePostHog, type EventName } from '@/hooks/posthog'
-import { getAppleAuthorizeURL } from '@/utils/auth'
+import { useToast } from '@/components/Toast/use-toast'
 import Apple from '@mui/icons-material/Apple'
 import { schemas } from '@polar-sh/client'
 import Button, { type ButtonProps } from '@polar-sh/ui/components/atoms/Button'
-import Link from 'next/link'
+import { getAppleAuthorizeURL } from '@/utils/auth'
 
 interface AppleLoginButtonProps {
+  authenticationSession: schemas['AuthenticationSession'] | null
   returnTo?: string
-  signup?: schemas['UserSignupAttribution']
+  signup?: boolean
   variant?: ButtonProps['variant']
 }
 
 const AppleLoginButton = ({
+  authenticationSession,
   returnTo,
   signup,
   variant,
 }: AppleLoginButtonProps) => {
   const posthog = usePostHog()
+  const authSessionStart = useAuthSessionStart()
+  const { toast } = useToast()
 
-  const onClick = () => {
+  const onClick = async () => {
     let eventName: EventName = 'global:user:login:submit'
     if (signup) {
       eventName = 'global:user:signup:submit'
     }
-
     posthog.capture(eventName, {
       method: 'apple',
     })
+
+    if (!authenticationSession) {
+      try {
+        await authSessionStart.mutateAsync(returnTo)
+      } catch {
+        toast({
+          title: 'Sign in failed',
+          description:
+            'Could not start authentication session. Please try again.',
+        })
+        return
+      }
+    }
+    window.location.href = getAppleAuthorizeURL()
   }
 
   return (
-    <Link
-      href={getAppleAuthorizeURL({
-        return_to: returnTo,
-        attribution: JSON.stringify(signup),
-      })}
-      onClick={onClick}
-    >
+    <a onClick={onClick}>
       <Button
         variant={variant}
         wrapperClassNames="space-x-2 p-2.5 px-5"
@@ -47,7 +59,7 @@ const AppleLoginButton = ({
           {signup ? 'Sign up with Apple' : 'Sign in with Apple'}
         </div>
       </Button>
-    </Link>
+    </a>
   )
 }
