+++ clients/apps/web/src/components/Auth/GoogleLoginButton.tsx
@@ -1,42 +1,54 @@
+import { useAuthSessionStart } from '@/hooks'
 import { usePostHog, type EventName } from '@/hooks/posthog'
-import { getGoogleAuthorizeLoginURL } from '@/utils/auth'
+import { useToast } from '@/components/Toast/use-toast'
 import Google from '@mui/icons-material/Google'
 import { schemas } from '@polar-sh/client'
 import Button, { type ButtonProps } from '@polar-sh/ui/components/atoms/Button'
-import Link from 'next/link'
+import { getGoogleAuthorizeLoginURL } from '@/utils/auth'
 
 interface GoogleLoginButtonProps {
+  authenticationSession: schemas['AuthenticationSession'] | null
   returnTo?: string
-  signup?: schemas['UserSignupAttribution']
+  signup?: boolean
   variant?: ButtonProps['variant']
 }
 
 const GoogleLoginButton = ({
+  authenticationSession,
   returnTo,
   signup,
   variant,
 }: GoogleLoginButtonProps) => {
   const posthog = usePostHog()
+  const authSessionStart = useAuthSessionStart()
+  const { toast } = useToast()
 
-  const onClick = () => {
+  const onClick = async () => {
     let eventName: EventName = 'global:user:login:submit'
     if (signup) {
       eventName = 'global:user:signup:submit'
     }
-
     posthog.capture(eventName, {
       method: 'google',
     })
+
+    if (!authenticationSession) {
+      try {
+        await authSessionStart.mutateAsync(returnTo)
+      } catch {
+        toast({
+          title: 'Sign in failed',
+          description:
+            'Could not start authentication session. Please try again.',
+        })
+        return
+      }
+    }
+    window.location.href = getGoogleAuthorizeLoginURL()
   }
 
   return (
-    <Link
-      href={getGoogleAuthorizeLoginURL({
-        return_to: returnTo,
-        attribution: JSON.stringify(signup),
-      })}
-      onClick={onClick}
-    >
+    <a onClick={onClick}>
       <Button
         variant={variant}
         wrapperClassNames="space-x-2 p-2.5 px-5"
@@ -47,7 +59,7 @@ const GoogleLoginButton = ({
           {signup ? 'Sign up with Google' : 'Sign in with Google'}
         </div>
       </Button>
-    </Link>
+    </a>
   )
 }
