--- clients/apps/web/src/components/Auth/LoginCodeForm.tsx
@@ -1,97 +1,0 @@
-'use client'
-
-import { LoginCodeError, useSendLoginCode } from '@/hooks/loginCode'
-import { usePostHog, type EventName } from '@/hooks/posthog'
-import { setValidationErrors } from '@/utils/api/errors'
-import { schemas } from '@polar-sh/client'
-import Button from '@polar-sh/ui/components/atoms/Button'
-import Input from '@polar-sh/ui/components/atoms/Input'
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormMessage,
-} from '@polar-sh/ui/components/ui/form'
-import { useState } from 'react'
-import { SubmitHandler, useForm } from 'react-hook-form'
-
-interface LoginCodeFormProps {
-  returnTo?: string
-  signup?: schemas['UserSignupAttribution']
-}
-
-const LoginCodeForm = ({ returnTo, signup }: LoginCodeFormProps) => {
-  const form = useForm<{ email: string }>({
-    defaultValues: {
-      email: '',
-    },
-  })
-  const { control, handleSubmit, setError } = form
-  const [loading, setLoading] = useState(false)
-  const sendLoginCode = useSendLoginCode()
-  const posthog = usePostHog()
-
-  const onSubmit: SubmitHandler<{ email: string }> = async ({ email }) => {
-    setLoading(true)
-    let eventName: EventName = 'global:user:login:submit'
-    if (signup) {
-      eventName = 'global:user:signup:submit'
-    }
-
-    posthog.capture(eventName, {
-      method: 'code',
-    })
-
-    try {
-      await sendLoginCode(email, returnTo, signup)
-    } catch (e) {
-      if (e instanceof LoginCodeError && e.error) {
-        setValidationErrors(e.error, setError)
-      }
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  return (
-    <Form {...form}>
-      <form className="flex w-full flex-col" onSubmit={handleSubmit(onSubmit)}>
-        <FormField
-          control={control}
-          name="email"
-          render={({ field }) => {
-            return (
-              <FormItem>
-                <FormControl className="w-full">
-                  <div className="flex w-full flex-col gap-2">
-                    <Input
-                      type="email"
-                      required
-                      placeholder="Email"
-                      autoComplete="off"
-                      data-1p-ignore
-                      {...field}
-                    />
-                    <Button
-                      type="submit"
-                      variant="secondary"
-                      fullWidth
-                      loading={loading}
-                      disabled={loading}
-                    >
-                      {signup ? 'Sign up with email' : 'Sign in with email'}
-                    </Button>
-                  </div>
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )
-          }}
-        />
-      </form>
-    </Form>
-  )
-}
-
-export default LoginCodeForm
