+++ clients/apps/web/src/components/Sandbox/SandboxBanner.tsx
@@ -4,12 +4,10 @@ import { usePathname } from 'next/navigation'
 
 const PUBLIC_PAGES = [
   '/',
-  '/login',
   '/auth',
   '/auth/email-otp',
   '/auth/totp',
   '/onboarding/sandbox',
-  '/login/code/verify',
 ]
 
 const SandboxBanner = () => {
